refactor: move mocks and fake-backend to another app
ref: #28
