+++ apps/front-end/back-office/src/app/services/auth/auth.service.spec.ts
@@ -7,10 +7,13 @@ import {
 } from '@angular/common/http/testing';
 import { beforeEach, describe, expect, it } from '@jest/globals';
 import { fail } from 'assert';
+import { authMocks } from '@webonjour/data-access-fake-backend';
 
 describe('AuthService', () => {
   let service: AuthService;
   let httpMock: HttpTestingController;
+  const response = authMocks.response;
+  const credentials = authMocks.credentials;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -109,14 +112,3 @@ describe('AuthService', () => {
     });
   });
 });
-
-export const credentials: Auth.LoginSchema = {
-  email: 'email@email.com',
-  password: 'password',
-};
-export const response: Auth.LoginResponse = {
-  accessToken:
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyLCJ0eXBlIjoiYWNjZXNzIn0.SWw75nUVldk2qUs1wiTx7ZIpjPAMLuGwOnxs_NMVUJM',
-  refreshToken:
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyLCJ0eXBlIjoicmVmcmVzaCJ9.rzQhdqE2VpiJ8mdD24FghBZ71D64CWMwQR5_c3_z-w0',
-};
