+++ libs/data-access-fake-backend/README.md
@@ -1,0 +1,7 @@
+# data-access-fake-backend
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test data-access-fake-backend` to execute the unit tests.
