+++ libs/data-access-fake-backend/project.json
@@ -1,0 +1,34 @@
+{
+  "name": "data-access-fake-backend",
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/data-access-fake-backend/src",
+  "prefix": "webonjour",
+  "targets": {
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/data-access-fake-backend/jest.config.ts",
+        "passWithNoTests": true
+      },
+      "configurations": {
+        "ci": {
+          "ci": true,
+          "codeCoverage": true
+        }
+      }
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/data-access-fake-backend/**/*.ts",
+          "libs/data-access-fake-backend/**/*.html"
+        ]
+      }
+    }
+  },
+  "tags": []
+}
