+++ libs/data-access-fake-backend/src/index.ts
@@ -1,0 +1,3 @@
+export { fakeBackendProvider } from './lib/interceptors/fake-backend.interceptor';
+// export all mocks in mocks folder
+export * as authMocks from './lib/mocks/auth';
