+++ tsconfig.base.json
@@ -15,6 +15,9 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
+      "@webonjour/data-access-fake-backend": [
+        "libs/data-access-fake-backend/src/index.ts"
+      ],
       "@webonjour/front-end/shared/common": [
         "libs/front-end/shared/common/src/index.ts"
       ],
